Don't copy old backups to /tmp
This patch will filter out the copying of old backups before they reach /tmp. If the backups folder isn't in the userdata folder, then the argument will be ignored by find, and there are no bad effects.

On my system, where the backups are in userdata, it was copying GBs of old backups to `/tmp` and then deleting them. `/tmp` is often in memory or quite small, so this should be avoided. It also speeds up my backups significantly.

Signed-off-by: James Hewitt-Thomas <james.hewitt@gmail.com>